day23/py/pyproject.toml
[tool.poetry]
name = "day23"
version = "0.1.0"
description = ""
authors = ["Saul Reynolds-Haertle <sreynoldshaertle@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.5"
more-itertools = "^8.12.0"
rich = "^10.16.1"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
ipython = "^7.30.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
part1 = "day23.part1:main"
part2 = "day23.part2:main"
